De-duplicate SetSpanOperation class
